Factor out key processor into testable package.
