+++ Packages/README.md
@@ -1,0 +1,7 @@
+# Packages
+
+Packages used to build this app.
+
+## KeyProcessor
+
+State machine that handles key press events. Factored out so it can be tested.
