+++ Packages/Tests/KeyProcessorTests/KeyProcessorTests.swift
@@ -1,0 +1,7 @@
+import XCTest
+@testable import KeyProcessor
+
+final class KeyProcessorTests: XCTestCase {
+    func testExample() {
+    }
+}
